weekly-ai-podcast/requirements.txt
anthropic>=0.39.0
elevenlabs>=1.0.0
pydub>=0.25.1
pyyaml>=6.0.1
python-dotenv>=1.0.1
requests>=2.32.0
pytest>=8.0.0
